Refactory + feature: change method argument + search for roe sentences
Now candidate_sentences work with a list of target_sets reather just one
target_set. This will help in future refactories linked with filtering
phase.

Searching for roe sentences. Dont making value searcher yet because of
the refactory that will be done this week and will change this part to
work with lists. And I know that the reports until now dont have this
data to be found.